Update available (#17)
* adds functionality to emit event when service worker is updated.

This allows us to prompt users to reload latest service worker content. As part of this process, we hand complete control to the user via skipWaiting and clientsClaim.

* updates depenencies and scripts

This updates dependencies and scripts to include http-server for a quasi-staging environment, serving production assets. Also gets rid of unused axios dependency.

* adds snackbar to prompt reload when new service worker is available.

This adds a snackbar, allowing the user to reload the page with latest service worker content.

* removes unused getter

* resolves lint error