Notification updates (#20)
* adds push notification for default state

On initial Wrapper invocation, the code checks for notification permissions. If no permissions are defined (ie. permissions are 'default'), display a sample notification. The notification serves the purpose of informing the user of notification location while also prompting for permission.

* push notification enhancements for Wrapper

This humanizes the expiration timer message and reoders imports so absolute imports come before relative.

* adds Howler as a dependency

Adds Howler as a basis for alert chime so DownTimer has a more robust audio system.

* applies Howler in alert_chime plugin

Applies Howler as the alert_chime platform, abstracting the constructor call via the plugin.

* converts notifications to plugin

Converts notifications via Push.js to a plugin in order to abstract functionality.