fixes a duration bug when loading saved timers
When loading saved timers, the duration would not get set properly. This resulted in incorrect progress bars on initial load.

For example, the default duration is 20 minutes. If a single timer were saved with a duration of 30 minutes, the next page load would show a progress bar at 33.33% complete (30 duration, 20 remaining = 10/30).